Crud_neo4j/requirements.txt
fastapi
uvicorn[standard]
neo4j[async]
streamlit
pdfplumber
python-docx
python-multipart
langchain-unstructured
python-dotenv